headers: move fi_log.h/fi_prov.h -> include/rdma/providers
Specifically identify that these files are intended to be included by
providers.  They include interfaces that are covered by the libfabric
ABI guarantee, so keep them under include/rdma, but put them off in
their own subdirectory, just to be extra special specific.

Update everywhere in the code base that includes these files to
reflect the new locations.  In a few cases, remove spurrious includes
of these header files (e.g., if that location already includes <fi.h>,
which, in turn, includes <rdma/providers/fi_log.h> and
<rdma/providers/fi_prov.h>, or if that location does not use any of
the interfaces in fi_log.h or fi_prov.h).
