+++ prov/gni/include/gnix_util.h
@@ -1,6 +1,7 @@
 /*
  * Copyright (c) 2015 Los Alamos National Security, LLC. All rights reserved.
  * Copyright (c) 2015-2016 Cray Inc. All rights reserved.
+ * Copyright (c) 2016 Cisco Systems, Inc. All rights reserved.
  *
  * This software is available to you under a choice of one of two
  * licenses.  You may choose to be licensed under the terms of the GNU
@@ -40,7 +41,6 @@
 
 #include <stdio.h>
 #include <fi.h>
-#include "rdma/fi_log.h"
 
 extern struct fi_provider gnix_prov;
