+++ prov/verbs/src/ep_rdm/verbs_utils.h
@@ -1,5 +1,6 @@
 /*
  * Copyright (c) 2013-2015 Intel Corporation, Inc.  All rights reserved.
+ * Copyright (c) 2016 Cisco Systems, Inc. All rights reserved.
  *
  * This software is available to you under a choice of one of two
  * licenses.  You may choose to be licensed under the terms of the GNU
@@ -42,7 +43,6 @@
 
 #include <infiniband/verbs.h>
 
-#include <rdma/fi_log.h>
 #include "../fi_verbs.h"
 
 #if (defined(__ICC) || defined(__INTEL_COMPILER) ||	\
