+++ src/log.c
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2015, Cisco Systems, Inc. All rights reserved.
+ * Copyright (c) 2015-2016, Cisco Systems, Inc. All rights reserved.
  * Copyright (c) 2015, Intel Corp., Inc.  All rights reserved.
  *
  * This software is available to you under a choice of one of two
@@ -37,7 +37,6 @@
 #include <stdlib.h>
 
 #include <rdma/fi_errno.h>
-#include <rdma/fi_log.h>
 
 #include "fi.h"
